+++ NEWS.md
@@ -1,4 +1,9 @@
 
+# precisePlacement 0.2.0
+
+* Added repelAxisLabels function.
+* Update all code for linting and autoformatting
+
 # precisePlacement 0.1.0
 
-First release.
+* First release.